Remove unnecessary casts to short

+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/ColorTriggers/BGColorTrigger.cs
@@ -13,7 +13,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.ColorTrigg
         public override int ObjectID => (int)TriggerType.BG;
         
         /// <summary>The target Color ID of the trigger.</summary>
-        public new short TargetColorID => (short)(int)SpecialColorID.BG;
+        public new int TargetColorID => (int)SpecialColorID.BG;
 
         /// <summary>Initializes a new instance of the <seealso cref="BGColorTrigger"/> class.</summary>
         public BGColorTrigger() : base((int)SpecialColorID.BG) { }
@@ -21,7 +21,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.ColorTrigg
         /// <param name="duration">The duration of the trigger.</param>
         /// <param name="tintGround">The Tint Ground property of the trigger.</param>
         public BGColorTrigger(float duration, bool tintGround = false)
-            : base(duration, (short)(int)SpecialColorID.BG, false, tintGround) { }
+            : base(duration, (int)SpecialColorID.BG, false, tintGround) { }
 
         /// <summary>Returns a clone of this <seealso cref="BGColorTrigger"/>.</summary>
         public override GeneralObject Clone() => AddClonedInstanceInformation(new BGColorTrigger());
