+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/ColorTriggers/GRNDColorTrigger.cs
@@ -13,7 +13,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.ColorTrigg
         public override int ObjectID => (int)TriggerType.GRND;
         
         /// <summary>The target Color ID of the trigger.</summary>
-        public new short TargetColorID => (short)(int)SpecialColorID.GRND;
+        public new int TargetColorID => (int)SpecialColorID.GRND;
 
         /// <summary>Initializes a new instance of the <seealso cref="GRNDColorTrigger"/> class.</summary>
         public GRNDColorTrigger() : base((int)SpecialColorID.GRND) { }
@@ -21,7 +21,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.ColorTrigg
         /// <param name="duration">The duration of the trigger.</param>
         /// <param name="tintGround">The Tint Ground property of the trigger.</param>
         public GRNDColorTrigger(float duration, bool tintGround = false)
-            : base(duration, (short)(int)SpecialColorID.GRND, false, tintGround) { }
+            : base(duration, (int)SpecialColorID.GRND, false, tintGround) { }
 
         /// <summary>Returns a clone of this <seealso cref="GRNDColorTrigger"/>.</summary>
         public override GeneralObject Clone() => AddClonedInstanceInformation(new GRNDColorTrigger());
